pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "3dde"
version = "0.1.0"
description = "3D Diffusion Model for Physical Systems with Graph Structure"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Emilio", email = "emilio@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "memory-profiler>=0.61.0",
    "torch-geometric>=2.3.0",
    "networkx>=3.0",
    "h5py>=3.8.0",
    "pandas>=2.0.0",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
